+++ commands/System/reload.js
@@ -1,64 +1,42 @@
-exports.run = (client, msg, [type, name]) => {
+exports.run = async (client, msg, [type, name]) => {
+  let m;
+  let message;
   switch (type) {
     case "function":
-      msg.channel.sendMessage(`Attemping to reload function ${name}`).then((m) => {
-        client.funcs.reload.function(client, client.clientBaseDir, name).then((message) => {
-          m.edit(`:white_check_mark: ${message}`);
-        }).catch((e) => {
-          m.edit(`:x: ${e}`);
-        });
-      });
+      m = await msg.channel.sendMessage(`Attemping to reload function ${name}`).catch(err => client.funcs.log(err, "error"));
+      message = await client.funcs.reload.function(client, client.clientBaseDir, name).catch(response => m.edit(`:x: ${response}`));
+      m.edit(`:white_check_mark: ${message}`);
       break;
     case "inhibitor":
-      msg.channel.sendMessage(`Attempting to reload inhibitor ${name}`).then((m) => {
-        client.funcs.reload.inhibitor(client, client.clientBaseDir, name).then((message) => {
-          m.edit(`:white_check_mark: ${message}`);
-        }).catch((e) => {
-          m.edit(`:x: ${e}`);
-        });
-      });
+      m = await msg.channel.sendMessage(`Attempting to reload inhibitor ${name}`).catch(err => client.funcs.log(err, "error"));
+      message = await client.funcs.reload.inhibitor(client, client.clientBaseDir, name).catch(response => m.edit(`:x: ${response}`));
+      m.edit(`:white_check_mark ${message}`);
       break;
     case "monitor":
-      msg.channel.sendMessage(`Attempting to reload monitor ${name}`).then((m) => {
-        client.funcs.reload.monitor(client, client.clientBaseDir, name).then((message) => {
-          m.edit(`:white_check_mark: ${message}`);
-        }).catch((e) => {
-          m.edit(`:x: ${e}`);
-        });
-      });
+      m = await msg.channel.sendMessage(`Attempting to reload monitor ${name}`).catch(err => client.funcs.log(err, "error"));
+      message = await client.funcs.reload.monitor(client, client.clientBaseDir, name).catch(response => m.edit(`:x: ${response}`));
+      m.edit(`:white_check_mark: ${message}`);
       break;
     case "provider":
-      msg.channel.sendMessage(`Attempting to reload provider ${name}`).then((m) => {
-        client.funcs.reload.provider(client, client.clientBaseDir, name).then((message) => {
-          m.edit(`:white_check_mark: ${message}`);
-        }).catch((e) => {
-          m.edit(`:x: ${e}`);
-        });
-      });
+      m = await msg.channel.sendMessage(`Attempting to reload provider ${name}`).catch(err => client.funcs.log(err, "error"));
+      message = await client.funcs.reload.provider(client, client.clientBaseDir, name).catch(response => m.edit(`:x: ${response}`));
+      m.edit(`:white_check_mark: ${message}`);
       break;
     case "event":
-      msg.channel.sendMessage(`Attempting to reload event ${name}`).then((m) => {
-        client.funcs.reload.event(client, name).then((message) => {
-          m.edit(`:white_check_mark: ${message}`);
-        }).catch((e) => {
-          m.edit(`:x: ${e}`);
-        });
-      });
+      m = await msg.channel.sendMessage(`Attempting to reload event ${name}`).catch(err => client.funcs.log(err, "error"));
+      message = await client.funcs.reload.event(client, name).catch(response => m.edit(`:x: ${response}`));
+      m.edit(`:white_check_mark: ${message}`);
       break;
     case "command":
       switch (name) {
         case "all":
-          client.funcs.loadCommands(client);
+          await client.funcs.loadCommands(client);
           msg.channel.sendMessage(":white_check_mark: Reloaded all commands.");
           break;
         default:
-          msg.channel.sendMessage(`Attempting to reload command ${name}`).then((m) => {
-            client.funcs.reload.command(client, client.clientBaseDir, name).then((message) => {
-              m.edit(`:white_check_mark: ${message}`);
-            }).catch((e) => {
-              m.edit(`:x: ${e}`);
-            });
-          });
+          m = await msg.channel.sendMessage(`Attempting to reload command ${name}`).catch(err => client.funcs.log(err, "error"));
+          message = await client.funcs.reload.command(client, client.clientBaseDir, name).catch(response => m.edit(`:x: ${response}`));
+          m.edit(`:white_check_mark: ${message}`);
           break;
       }
       break;