+++ functions/loadSingleCommand.js
@@ -1,6 +1,6 @@
 const path = require("path");
 
-module.exports = (client, command, reload = false, loadPath = null) => new Promise((resolve, reject) => {
+module.exports = (client, command, reload = false, loadPath = null) => new Promise(async (resolve, reject) => {
   let category;
   let subCategory;
   let cmd;
@@ -41,14 +41,12 @@ module.exports = (client, command, reload = false, loadPath = null) => new Promi
     } catch (e) {
       if (e.code === "MODULE_NOT_FOUND") {
         const module = /'[^']+'/g.exec(e.toString());
-        client.funcs.installNPM(module[0].slice(1, -1))
-            .then(() => {
-              client.funcs.loadSingleCommand(client, command, false, loadPath);
-            })
+        await client.funcs.installNPM(module[0].slice(1, -1))
             .catch((err) => {
               console.error(err);
               process.exit();
             });
+        client.funcs.loadSingleCommand(client, command, false, loadPath);
       } else {
         reject(`Could not load new command data: ${e.stack}`);
       }