+++ CHANGELOG.md
@@ -8,11 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - New Beta Configuration (Needs heavy testing)
 - New Argument Prompting for Commands
+- New Initialize Function to aleviate undefined errors
 
 ### Changed
 - Confs.js uses new configuration system now
 - Configuration now split into smaller parts as requested.
 - Help command is now a Direct Message.
+- Async/Await for all pieces && app.js
+- dataProviders renamed to Providers
 
 ### Fixed
 - loadCommands no longer counts/loads "Ghost" commands.